twitter-app/backend/pyproject.toml
[tool.poetry]
name = "twitter-app"
package-mode = false
version = "0.1.0"
description = "app baseado no twitter"
authors = ["lmescoito <lmescoito@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.9"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
pip = "^24.1.2"
unicorn = "^2.0.1.post1"
faker = "^26.0.0"
factory-boy = "^3.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
